www/react: Extract WorkerBadge component to common ui package
